move json serialization package to subfolder
