--- serialization/serializer_test.go
@@ -1,13 +1,0 @@
-package serialization_test
-
-import (
-	"testing"
-
-	. "github.com/onsi/ginkgo/v2"
-	. "github.com/onsi/gomega"
-)
-
-func TestSerialization(t *testing.T) {
-	RegisterFailHandler(Fail)
-	RunSpecs(t, "Serialization Package")
-}
